Updated everything

+++ ClinicalSystem/src/main/java/com/example/ClinicalSystem/model/Appointment.java
@@ -16,7 +16,7 @@ public class Appointment {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	@Column(name = "date" , nullable = false)
 	private String date;
@@ -49,7 +49,7 @@ public class Appointment {
 		super();
 	}
 
-	public Appointment(int id, String date, String time, String type, Patient patient, Doctor doctor,
+	public Appointment(Long id, String date, String time, String type, Patient patient, Doctor doctor,
 			Appointment appointment, boolean hasHappend) {
 		super();
 		this.id = id;
@@ -62,11 +62,11 @@ public class Appointment {
 		this.hasHappend = hasHappend;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
